inclusive-language-detector-python/inclusive-language-detector-python/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
openai==1.30.1
gunicorn==21.2.0
python-dotenv==1.0.0
httpx==0.27.0